# openapi.yaml
openapi: 3.0.1
info:
  title: Delilah Resto
  description: Restaurant API.
  contact:
    email: fgbarrientos93@gmail.com
  version: 1.0.0
...
servers:
  - url: https://localhost:3000/
...
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get list of orders
      description: Returns a list of orders
      operationId: getOrders
      responses:
          200:
            description: "-"
          401:
            description: "You need to be logged"
          500:
            description: "An error with the database happened"        
    post:
      tags:
        - Orders
      summary: Adds a product
      description: Adds a product to database
      operationId: addOrder
      requestBody:
        description: Order to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewOrder"
        required: true
      responses:
        200: 
          description: "-"
        402:
          description: "You need to be logged"
        500:
         description: "An error with the database happened"            
  /orders/:id:
    get:
      tags:
        - Orders
      summary: Gets order by id
      parameters:
        - name: id
          in: path
          description: The ID of the order
          required: true
          schema:
            type: string
      description: Returns a specefic order
      operationId: getOrder
      responses:
        200: 
          description: "-"
        404:
          description: "The order you're looking for doesn't exist"
        500:
         description: "An error with the database happened"      

    put:
      tags:
        - Orders
      summary: Gets order by Id
      description: Edits a specific order
      operationId: editOrder
      parameters:
        - name: id
          in: path
          description: The ID of the order
          required: true
          schema:
            type: string
      requestBody:
        description: Order to create
        content:
          application/json:
            schema:
              type: object
              properties:
                order_status:
                  type: string
                  example: delivered
        required: true
      responses:
        200: 
          description: "-"
        404:
          description: "The order you're looking for doesn't exist"
        500:
         description: "An error with the database happened"      

    delete:
      tags:
        - Orders
      summary: Delete an order
      description: Detele an specific order
      parameters:
        - name: id
          in: path
          description: The ID of the order
          required: true
          schema:
            type: string
      operationId: deleteOrder
      responses:
        200:
          description: "order succesfully deleted"
        404:
          description: "Order not found"
        500:
          description: "An error with the database happened"  
components:
  schemas:
...
    NewOrder:
      type: object
      properties:
        payment_method:
          type: string
          example: cash
        products:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
                example: 1
              amount:
                type: integer
                example: 3
security:
  - bearerAuth: []